refactor(lotto): Lotto Random 하드코딩된 문자열 제거
