+++ lotto/application/src/main/java/com/su8y/lotto/application/common/LottoConstant.java
@@ -1,0 +1,5 @@
+package com.su8y.lotto.application.common;
+
+public enum LottoConstant {
+	RANDOM
+}
