+++ lotto/application/src/main/java/com/su8y/lotto/application/port/in/command/LottoGenerationUseCase.java
@@ -1,6 +1,6 @@
 package com.su8y.lotto.application.port.in.command;
 
-import com.su8y.lotto.application.dto.LottoTicketResponse;
+import com.su8y.lotto.application.common.LottoTicketResponse;
 import com.su8y.lotto.domain.LottoGenerator;
 
 public interface LottoGenerationUseCase {
