+++ lotto/application/src/main/java/com/su8y/lotto/application/service/LottoRandomGenerationCommand.java
@@ -1,5 +1,6 @@
 package com.su8y.lotto.application.service;
 
+import com.su8y.lotto.application.common.LottoConstant;
 import com.su8y.lotto.domain.LottoGenerator;
 
 import lombok.RequiredArgsConstructor;
@@ -9,7 +10,7 @@ public class LottoRandomGenerationCommand implements LottoGenerator.LottoGenerat
 	private final long userId;
 	@Override
 	public String getMethod() {
-		return "RANDOM";
+		return LottoConstant.RANDOM.name();
 	}
 	public Long getUserId(){
 		return this.userId;
