+++ lotto/application/src/main/java/com/su8y/lotto/application/service/LottoRandomGenerationMetadata.java
@@ -1,10 +1,11 @@
 package com.su8y.lotto.application.service;
 
+import com.su8y.lotto.application.common.LottoConstant;
 import com.su8y.lotto.domain.GenerationMetadata;
 
 public class LottoRandomGenerationMetadata implements GenerationMetadata {
 	@Override
 	public String getMethod() {
-		return "RANDOM";
+		return LottoConstant.RANDOM.name();
 	}
 }
